+++ src/models/ChainBlockData.ts
@@ -1,9 +1,17 @@
 import { BigNumber } from 'ethers';
 import { IChainInstance } from './ChainInstance';
 
+export interface ChainFCDData {
+  dataTimestamp: BigNumber;
+  value: string;
+}
+
+export type ChainFCDsData = [timestamps: BigNumber[], values: BigNumber[]];
+
 export interface ChainBlockData {
   anchor: BigNumber;
-  timestamp: BigNumber;
+  dataTimestamp: BigNumber;
+  timestamp: number;
   root: string;
   minter: string;
   staked: BigNumber;