+++ CHANGELOG.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
 - added statsd support and reporting of controller requests
+
+### Changed
+- support storing only latest FCDs
+
+### Fixed
+- do not mark the latest block as failed right away, when can't get leaves from validators, retry when it happens
 
 ## [0.3.2] - 2021-05-03
 ### Fixed